--- internal/ctxval/ctx_keys.go
@@ -1,10 +1,0 @@
-package ctxval
-
-type commonKeyId int
-
-const (
-	loggerCtxKey     commonKeyId = iota
-	requestIdCtxKey  commonKeyId = iota
-	requestNumCtxKey commonKeyId = iota
-	accountCtxKey    commonKeyId = iota
-)
